encapsulate getTxFeeHandler for serial usage
