--- app/app_partial_concurrent.go
@@ -1,45 +1,0 @@
-package app
-
-import (
-	"encoding/hex"
-	"strings"
-
-	sdk "github.com/okex/exchain/libs/cosmos-sdk/types"
-	"github.com/okex/exchain/libs/cosmos-sdk/x/auth"
-	authante "github.com/okex/exchain/libs/cosmos-sdk/x/auth/ante"
-	evmtypes "github.com/okex/exchain/x/evm/types"
-)
-
-// getTxFeeAndFromHandler get tx fee and from
-func getTxFeeAndFromHandler(ak auth.AccountKeeper) sdk.GetTxFeeAndFromHandler {
-	return func(ctx sdk.Context, tx sdk.Tx, justFee bool) (fee sdk.Coins, isEvm bool, from string, to string, err error) {
-		if evmTx, ok := tx.(*evmtypes.MsgEthereumTx); ok {
-			isEvm = true
-			err = evmTx.VerifySig(evmTx.ChainID(), ctx.BlockHeight())
-			if err != nil {
-				return
-			}
-			fee = evmTx.GetFee()
-			if justFee {
-				return
-			}
-			from = evmTx.BaseTx.From
-			if len(from) > 2 {
-				from = strings.ToLower(from[2:])
-			}
-			if evmTx.To() != nil {
-				to = strings.ToLower(evmTx.To().String()[2:])
-			}
-		} else if feeTx, ok := tx.(authante.FeeTx); ok {
-			fee = feeTx.GetFee()
-			if justFee {
-				return
-			}
-			feePayer := feeTx.FeePayer(ctx)
-			feePayerAcc := ak.GetAccount(ctx, feePayer)
-			from = hex.EncodeToString(feePayerAcc.GetAddress())
-		}
-
-		return
-	}
-}
