+++ libs/cosmos-sdk/baseapp/baseapp.go
@@ -149,6 +149,7 @@ type BaseApp struct { // nolint: maligned
 	logFix                       sdk.LogFix
 
 	getTxFeeAndFromHandler sdk.GetTxFeeAndFromHandler
+	getTxFeeHandler sdk.GetTxFeeHandler
 
 	// volatile states:
 	//
