+++ libs/cosmos-sdk/baseapp/baseapp_runtx.go
@@ -125,8 +125,8 @@ func (app *BaseApp) runtxWithInfo(info *runTxInfo, mode runTxMode, txBytes []byt
 	}
 	app.pin(RunAnte, false, mode)
 
-	if app.getTxFeeAndFromHandler != nil && mode == runTxModeDeliver {
-		fee, _, _, _, _ := app.getTxFeeAndFromHandler(info.ctx, tx, true)
+	if app.getTxFeeHandler != nil && mode == runTxModeDeliver {
+		fee, _ := app.getTxFeeHandler(info.ctx, tx)
 		app.UpdateFeeForCollector(fee, true)
 	}
