+++ libs/cosmos-sdk/types/handler.go
@@ -17,7 +17,8 @@ type UpdateFeeCollectorAccHandler func(ctx Context, balance Coins) error
 
 type LogFix func(logIndex []int, errs []error) (logs [][]byte)
 
-type GetTxFeeAndFromHandler func(ctx Context, tx Tx, justFee bool) (Coins, bool, string, string, error)
+type GetTxFeeAndFromHandler func(ctx Context, tx Tx) (Coins, bool, string, string, error)
+type GetTxFeeHandler func(ctx Context, tx Tx) (Coins, error)
 
 // AnteDecorator wraps the next AnteHandler to perform custom pre- and post-processing.
 type AnteDecorator interface {
